q1_sports_analytics_rag/requirements.txt
openai>=1.12.0,<2.0.0
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-community>=0.0.20,<0.1.0
langchain-core>=0.1.0,<0.2.0
langchain-text-splitters>=0.0.1,<0.1.0
chromadb>=0.4.22,<0.5.0
sentence-transformers>=2.2.2,<3.0.0
fastapi>=0.104.1,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.5.0,<3.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
tiktoken==0.5.2